requirements.txt
streamlit
groq
python-dotenv
ollama